feat: add cart modal
I added a modal cart with product number and final value after deciding
how many products. This was a harder job than I thought, but I did, I put the whole app in a context
because I needed the variables 'count' and 'finalValue' to be used in
all application, I was able to understand a very useful new concept and
now, mine focus will be on updating the components for a better understanding.

I added a modal for cart products with shadcn-ui and radix-ui.